Implements Solid.surface_area
The Solid class now has a surface_area property which is a
scalar/unitless number (since STL files are inherently unitless) that
corresponds to the calculated surface area of the object.

The Solid's surface area is simply a summation of the area of each
facet.

WIP surface area

Signed-off-by: zachwick <zach@zachwick.com>

Returns Facet.area now - d'oh

Signed-off-by: zachwick <zach@zachwick.com>